Fixed status bar style hidden by overlay window
Introduced a view controller for the overlay window to preserve the preferred status bar style for the main window. This prevents losing the status bar style when the new window is created and its own status bar hides the main window.

One way to reproduce this issue is by adding a navigation bar and changing its style to be dark (with white fonts) and then tap anywhere.

Originally the status bar was going to change its appearance to the default style.
Currently is going to be set based on the keyWIndow rootViewController preferredStatusBarStyle property.